+++ valhalla/filters/collection.py
@@ -1,3 +1,3 @@
-def pluck_keys(): pass # removes keys from a dictionary
+def drop_keys(): pass # removes keys from a dictionary
 
 def contains(): pass # tests to make sure that the input value is within the specified list