+++ CHANGELOG.md
@@ -8,6 +8,7 @@ current version: 0.0.3
 * Added tests for ```strings.canonize``` and ```strings.slugify```
 * Fixed the ```web.ipv4``` validator so that it actually verifies the values of the octets :0
 * Added the following filters: ```[casting.boolean, casting.jsbool, casting.strbool, casting.integer, casting.longint, casting.numeric]```
-* filters continued: ```[web.url]```
+* filters continued: ```[web.url, casting.string, chrono.date, chrono.time, chrono.datetime, chrono.time_before, chrono.time_after, chrono.time_between]```
 * Added a schema-level ```match``` option used like so: ```s = Schema('my_schema', match={'password': 'password_confirm'})```
 * Changed the name of the ```strings.numeric``` validator to ```strings.numeric_string``` to differentiate it from the ```casting.numeric``` filter. Remember that filters are thrust into a global namespace as part of the API's brevity.
+* Added tests for the chrono module